Added interfaces and implements service
