+++ src/main/java/br/com/grupocesw/easyong/services/JwtTokenService.java
@@ -3,137 +3,33 @@ package br.com.grupocesw.easyong.services;
 import java.io.IOException;
 import java.util.Date;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Service;
-import org.springframework.web.filter.OncePerRequestFilter;
 
-import br.com.grupocesw.easyong.configs.JwtConfig;
-import br.com.grupocesw.easyong.entities.UserAuthDetails;
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.SignatureException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import lombok.extern.slf4j.Slf4j;
 
 @Service
-@Slf4j
-public class JwtTokenService extends OncePerRequestFilter {
+public interface JwtTokenService {
 
-	@Autowired private JwtConfig jwtConfig;
-    @Autowired private UserService userService;
+    public String generateToken(Authentication authentication);
 
-    public String generateToken(Authentication authentication) {
-
-        Long now = System.currentTimeMillis();
-        return Jwts.builder()
-                .setSubject(authentication.getName())
-                .claim("authorities", authentication.getAuthorities().stream()
-                        .map(GrantedAuthority::getAuthority).collect(Collectors.toList()))
-                .setIssuedAt(new Date(now))
-                .setExpiration(new Date(now + jwtConfig.getExpiration() * 1000))  // in milliseconds
-                .signWith(SignatureAlgorithm.HS512, jwtConfig.getSecret().getBytes())
-                .compact();
-    }
-
-    public Claims getClaimsFromJWT(String token) {
-        return Jwts.parser()
-                .setSigningKey(jwtConfig.getSecret().getBytes())
-                .parseClaimsJws(token)
-                .getBody();
-    }
+    public Claims getClaimsFromJWT(String token);
     
-    private Claims getAllClaimsFromToken(String token) {
-    	return Jwts.parser().setSigningKey(jwtConfig.getSecret()).parseClaimsJws(token).getBody();
-    }
+    public Claims getAllClaimsFromToken(String token);
     
-    public <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver) {
-    	final Claims claims = getAllClaimsFromToken(token);
-    	return claimsResolver.apply(claims);
-
-    }
+    public <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver);
     
-    public Date getExpirationDateFromToken(String token) {
-    	return getClaimFromToken(token, Claims::getExpiration);
-    }
+    public Date getExpirationDateFromToken(String token);
     
-    @SuppressWarnings("unused")
-	private Boolean isTokenExpired(String token) {
-    	final Date expiration = getExpirationDateFromToken(token);
-    	return expiration.before(new Date());
-    }
+    public Boolean isTokenExpired(String token);
 
-    public boolean validateToken(String authToken) {
-        try {
-            Jwts.parser()
-                    .setSigningKey(jwtConfig.getSecret().getBytes())
-                    .parseClaimsJws(authToken);
+    public boolean validateToken(String authToken);
 
-            return true;
-        } catch (SignatureException ex) {
-            log.error("Invalid JWT signature");
-        } catch (MalformedJwtException ex) {
-            log.error("Invalid JWT token");
-        } catch (ExpiredJwtException ex) {
-            log.error("Expired JWT token");
-        } catch (UnsupportedJwtException ex) {
-            log.error("Unsupported JWT token");
-        } catch (IllegalArgumentException ex) {
-            log.error("JWT claims string is empty.");
-        }
-        return false;
-    }
-    
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
-            throws ServletException, IOException {
-
-        String header = request.getHeader(jwtConfig.getHeader());
-
-        if (header == null || !header.startsWith(jwtConfig.getPrefix())) {
-            chain.doFilter(request, response);
-            return;
-        }
-
-        String token = header.replace(jwtConfig.getPrefix(), "");
-
-        if (this.validateToken(token)) {
-            Claims claims = this.getClaimsFromJWT(token);
-            String username = claims.getSubject();
-
-            UsernamePasswordAuthenticationToken auth =
-                    userService.findByUsername(username)
-                            .map(UserAuthDetails::new)
-                            .map(userDetails -> {
-                                UsernamePasswordAuthenticationToken authentication =
-                                        new UsernamePasswordAuthenticationToken(
-                                                userDetails, null, userDetails.getAuthorities());
-                                authentication
-                                        .setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-
-                                return authentication;
-                            })
-                            .orElse(null);
-
-            SecurityContextHolder.getContext().setAuthentication(auth);
-        } else {
-            SecurityContextHolder.clearContext();
-        }
-
-        chain.doFilter(request, response);
-    }
+    abstract void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws ServletException, IOException;
 }
