+++ src/main/java/br/com/grupocesw/easyong/services/NgoService.java
@@ -1,93 +1,23 @@
 package br.com.grupocesw.easyong.services;
 
-import java.util.Optional;
-
-import javax.persistence.EntityNotFoundException;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import br.com.grupocesw.easyong.entities.Ngo;
-import br.com.grupocesw.easyong.repositories.NgoRepository;
-import br.com.grupocesw.easyong.services.exceptions.DatabaseException;
-import br.com.grupocesw.easyong.services.exceptions.ResourceNotFoundException;
 
 @Service
-public class NgoService {
+public interface NgoService {
 
-	@Autowired
-	private NgoRepository repository;
+	public Page<Ngo> findByActivated(Pageable pageable);
 
-	public Page<Ngo> findByActivated(Pageable pageable) {
-		return repository.findByActivated(pageable);
-	}
+	public Page<Ngo> findSuggested(Pageable pageable);
+
+	public Ngo create(Ngo ngo);
 	
-	public Page<Ngo> findSuggested(Pageable pageable) {
-		return repository.findSuggested(pageable);
-	}
+	public Ngo retrieve(Long id);
 
-	public Ngo findById(Long id) {
-		Optional<Ngo> optional = repository.findById(id);
+	public Ngo update(Long id, Ngo payload) throws Exception;
 
-		return optional.orElseThrow(() -> new ResourceNotFoundException(id));
-	}
-
-	public Ngo insert(Ngo ngo) {		
-		try {
-			return repository.save(ngo);
-		} catch (DataIntegrityViolationException e) {
-			throw new DatabaseException(e.getMessage());
-		}
-	}
-
-	public Ngo update(Long id, Ngo payload) throws Exception {
-		try {
-			Ngo ngo = repository.getOne(id);
-			this.updateData(ngo, payload);
-
-			return repository.save(ngo);
-		} catch (EntityNotFoundException e) {
-			throw new ResourceNotFoundException(id);
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
-	}
-
-	private void updateData(Ngo ngo, Ngo payload) {
-		ngo.setName(payload.getName());
-		ngo.setCnpj(payload.getCnpj());
-		ngo.setDescription(payload.getDescription());
-		ngo.setActivated(payload.getActivated());
-		
-		if (payload.getAddress() != null) {
-			ngo.getAddress().setNumber(payload.getAddress().getNumber());
-			ngo.getAddress().setComplement(payload.getAddress().getComplement());
-			ngo.getAddress().setLatitude(payload.getAddress().getLatitude());
-			ngo.getAddress().setLongitude(payload.getAddress().getLongitude());
-			
-			if (payload.getAddress().getStreet() != null) {
-				ngo.getAddress().getStreet().setId(payload.getAddress().getStreet().getId());
-			}
-		}
-		
-		if (payload.getCauses() != null) {
-			ngo.getCauses().clear();
-			ngo.getCauses().addAll(payload.getCauses());
-		}			
-	}
-
-	public void delete(Long id) {
-		try {
-			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
-		} catch (DataIntegrityViolationException e) {
-			throw new DatabaseException(e.getMessage());
-		}
-	}
-
+	public void delete(Long id);
 }
