+++ src/main/java/br/com/grupocesw/easyong/services/impl/JwtTokenServiceImpl.java
@@ -1,0 +1,146 @@
+package br.com.grupocesw.easyong.services.impl;
+
+import java.io.IOException;
+import java.util.Date;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Service;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import br.com.grupocesw.easyong.configs.JwtConfig;
+import br.com.grupocesw.easyong.entities.UserAuthDetails;
+import br.com.grupocesw.easyong.services.JwtTokenService;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.SignatureException;
+import io.jsonwebtoken.UnsupportedJwtException;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@Slf4j
+public class JwtTokenServiceImpl extends OncePerRequestFilter implements JwtTokenService {
+
+	@Autowired private JwtConfig jwtConfig;
+    @Autowired private UserServiceImpl userService;
+
+    @Override
+    public String generateToken(Authentication authentication) {
+
+        Long now = System.currentTimeMillis();
+        return Jwts.builder()
+                .setSubject(authentication.getName())
+                .claim("authorities", authentication.getAuthorities().stream()
+                        .map(GrantedAuthority::getAuthority).collect(Collectors.toList()))
+                .setIssuedAt(new Date(now))
+                .setExpiration(new Date(now + jwtConfig.getExpiration() * 1000))  // in milliseconds
+                .signWith(SignatureAlgorithm.HS512, jwtConfig.getSecret().getBytes())
+                .compact();
+    }
+
+    @Override
+    public Claims getClaimsFromJWT(String token) {
+        return Jwts.parser()
+                .setSigningKey(jwtConfig.getSecret().getBytes())
+                .parseClaimsJws(token)
+                .getBody();
+    }
+    
+    @Override
+    public Claims getAllClaimsFromToken(String token) {
+    	return Jwts.parser().setSigningKey(jwtConfig.getSecret()).parseClaimsJws(token).getBody();
+    }
+    
+    @Override
+    public <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver) {
+    	final Claims claims = getAllClaimsFromToken(token);
+    	return claimsResolver.apply(claims);
+
+    }
+    
+    @Override
+    public Date getExpirationDateFromToken(String token) {
+    	return getClaimFromToken(token, Claims::getExpiration);
+    }
+    
+    @Override
+	public Boolean isTokenExpired(String token) {
+    	final Date expiration = getExpirationDateFromToken(token);
+    	return expiration.before(new Date());
+    }
+
+    @Override
+    public boolean validateToken(String authToken) {
+        try {
+            Jwts.parser()
+                    .setSigningKey(jwtConfig.getSecret().getBytes())
+                    .parseClaimsJws(authToken);
+
+            return true;
+        } catch (SignatureException ex) {
+            log.error("Invalid JWT signature");
+        } catch (MalformedJwtException ex) {
+            log.error("Invalid JWT token");
+        } catch (ExpiredJwtException ex) {
+            log.error("Expired JWT token");
+        } catch (UnsupportedJwtException ex) {
+            log.error("Unsupported JWT token");
+        } catch (IllegalArgumentException ex) {
+            log.error("JWT claims string is empty.");
+        }
+        return false;
+    }
+    
+    @Override
+    public void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
+            throws ServletException, IOException {
+
+        String header = request.getHeader(jwtConfig.getHeader());
+
+        if (header == null || !header.startsWith(jwtConfig.getPrefix())) {
+            chain.doFilter(request, response);
+            return;
+        }
+
+        String token = header.replace(jwtConfig.getPrefix(), "");
+
+        if (this.validateToken(token)) {
+            Claims claims = this.getClaimsFromJWT(token);
+            String username = claims.getSubject();
+
+            UsernamePasswordAuthenticationToken auth =
+                    userService.findByUsername(username)
+                            .map(UserAuthDetails::new)
+                            .map(userDetails -> {
+                                UsernamePasswordAuthenticationToken authentication =
+                                        new UsernamePasswordAuthenticationToken(
+                                                userDetails, null, userDetails.getAuthorities());
+                                authentication
+                                        .setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+
+                                return authentication;
+                            })
+                            .orElse(null);
+
+            SecurityContextHolder.getContext().setAuthentication(auth);
+        } else {
+            SecurityContextHolder.clearContext();
+        }
+
+        chain.doFilter(request, response);
+    }
+}
