+++ src/main/java/br/com/grupocesw/easyong/services/impl/UserAuthDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package br.com.grupocesw.easyong.services;
+package br.com.grupocesw.easyong.services.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Service;
 import br.com.grupocesw.easyong.entities.UserAuthDetails;
 
 @Service
-public class UserAuthDetailsService implements UserDetailsService {
+public class UserAuthDetailsServiceImpl implements UserDetailsService {
 
     @Autowired
-    private UserService userService;
+    private UserServiceImpl userService;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
