+++ src/main/resources/application-staging.properties
@@ -1,16 +1,12 @@
-server.port=8080
+spring.datasource.url=jdbc:postgresql://localhost:5432/easyong
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.datasource.data=classpath:/database/initial-data.sql, classpath:/database/initial-data-dev.sql, classpath:/database/update-sequence.sql
 
-spring.datasource.url=jdbc:h2:mem:easyong
-spring.datasource.platform=h2
-spring.datasource.username=sa
-spring.datasource.password=
-spring.datasource.data=classpath:/database/initial-data.sql, classpath:/database/initial-data-dev.sql
-
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+spring.datasource.initialization-mode=always
 
 # DATA WEB (SpringDataWebProperties)
 spring.data.web.pageable.default-page-size=10
@@ -24,8 +20,8 @@ spring.data.web.sort.sort-parameter=sort
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-#spring.mail.username=grupocesw@gmail.com
-#spring.mail.password=grupocesw@2020
+spring.mail.username=grupocesw@gmail.com
+spring.mail.password=grupocesw@2020
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.connectiontimeout=5000
 spring.mail.properties.mail.smtp.timeout=5000
